plist de producto

+++ src/app/component/producto/producto-routed-plist/producto-routed-plist.component.ts
@@ -4,10 +4,10 @@ import { ProductoService } from 'src/app/service/producto.service';
 
 import { Router, ActivatedRoute, ParamMap } from '@angular/router';
 import { Location } from '@angular/common';
-import { MatSort } from '@angular/material/sort';
+
 import { MatPaginator, PageEvent } from '@angular/material/paginator';
 import { MatTableDataSource } from '@angular/material/table';
-import { Sort } from '@angular/material/sort';
+
 import { MatSnackBar } from '@angular/material/snack-bar';
 
 @Component({
@@ -17,16 +17,22 @@ import { MatSnackBar } from '@angular/material/snack-bar';
 })
 
 export class ProductoRoutedPlistComponent implements OnInit {
-  currentSort: Sort | undefined = undefined;
+  entity: String = "producto";
+
+  currentPageIndex: number = 0;
+  currentPageSize: number = 10;
+  currentSortField: string = "id";
+  currentSortDirection: string = "asc";
+  currentFilter: string = "";
+
   totalElements: number = 0;
   totalPages: number = 0;
-  pageIndex: number = 0;
-  pageSize: number = 10;
   page: IPage = new Page();
   Products: IProducto[] | undefined;
-  columnas: string[] = ['id', 'nombre', 'codigo', 'precio', 'existencias', 'tipo', 'herramientas'];
+
   loading: boolean = false;
-  oRouter: Router | undefined;
+
+  paginationPad: string[] = [];
 
   //new MatTableDataSource(<IProducto> response);
   dataSource: MatTableDataSource<IProducto> = new MatTableDataSource();
@@ -34,35 +40,85 @@ export class ProductoRoutedPlistComponent implements OnInit {
   @ViewChild(MatPaginator) paginator: MatPaginator | undefined;
   //@ViewChild(MatSort) sort: MatSort | undefined;
 
-  getPage = (pageNumber: number, rpp: number, sort: Sort | undefined = undefined, filter: string | undefined = undefined) => {
+  pagination = (pages: number, page: number) => {
+    console.log("pag", pages, page);
+    let botonera: string[] = [];
+    let i: number;
+    let neighbourhood: number = 2;
+    for (i = 1; i <= pages; i++) {
+      if (i == 1) {
+        botonera.push(i.toString());
+      } else if (i > (page - neighbourhood) && i < (page + neighbourhood)) {
+        botonera.push(i.toString());
+      } else if (i == pages) {
+        botonera.push(i.toString());
+      } else if (i == (page - neighbourhood) || i == (page + neighbourhood)) {
+        botonera.push('...');
+      }
+    }
+    return botonera;
+  }
+
+  getPage = (pageNumber: number, rpp: number, sortField: string, sortDirection: string, filter: string | undefined = undefined) => {
+    this._snackBar.open("Cargando datos","Por favor espera...");
     this.loading = true;
-    this.productoService.getProductos(pageNumber, rpp, sort, filter).subscribe((oPage: IPage) => {
-      this.page = oPage;
-      this.Products = oPage.content;
-      this.totalElements = oPage['totalElements'];
-      this.totalPages = oPage['totalPages'];
-      this.dataSource = new MatTableDataSource(oPage.content);
-      //console.log(this.dataSource);   
-      this.pageSize = oPage['size'];
-      this.pageIndex = pageNumber;
+    this.productoService.getProductosBootstrap(pageNumber - 1, rpp, sortField, sortDirection, filter).subscribe((oPage: IPage) => {
+      if (pageNumber > oPage['totalPages']) {
+        pageNumber = oPage['totalPages'];
+        this.productoService.getProductosBootstrap(pageNumber - 1, rpp, sortField, sortDirection, filter).subscribe((oPage: IPage) => {
+          this.page = oPage;
+          this.Products = oPage.content;
+          this.totalElements = oPage['totalElements'];
+          this.totalPages = oPage['totalPages'];
+          this.paginationPad = this.pagination(this.totalPages, this.currentPageIndex);
+          this.currentPageSize = oPage['size'];
+          this.currentPageIndex = pageNumber;
+        }, error => {
+          if (error.status == 401) {
+            //se ha perdido la sesión        
+            this.openSnackBar("Su sesión ha expirado. Por favor debe volver a autenticarse en la página de login.", "ERROR DE SESIÓN");
+            localStorage.clear();
+            if (this.oRouter !== undefined) {
+              this.oRouter.navigate(['/login']);
+            }
+          } else {
+            //servidor caido
+            this.openSnackBar("Hay algún problema técnico con el servidor en Internet. Por favor avise al técnico de la aplicación.", "ERROR DE SERVIDOR");
+            localStorage.clear();
+            if (this.oRouter !== undefined) {
+              this.oRouter.navigate(['/home']);
+            }
+          }
+        })
+      } else {
+        this.page = oPage;
+        this.Products = oPage.content;
+        this.totalElements = oPage['totalElements'];
+        this.totalPages = oPage['totalPages'];
+        this.paginationPad = this.pagination(this.totalPages, this.currentPageIndex);
+        this.currentPageSize = oPage['size'];
+        this.currentPageIndex = pageNumber;
+      }
       this.loading = false;
+      this._snackBar.dismiss();
     }, error => {
       if (error.status == 401) {
         //se ha perdido la sesión        
         this.openSnackBar("Su sesión ha expirado. Por favor debe volver a autenticarse en la página de login.", "ERROR DE SESIÓN");
-        localStorage.clear(); 
+        localStorage.clear();
         if (this.oRouter !== undefined) {
           this.oRouter.navigate(['/login']);
         }
       } else {
         //servidor caido
-        this.openSnackBar("Hay algún problema técnico con el servidor en Internet. Por favor avise al técnico de la aplicación.", "ERROR DE SERVIDOR");        
-        localStorage.clear(); 
+        this.openSnackBar("Hay algún problema técnico con el servidor en Internet. Por favor avise al técnico de la aplicación.", "ERROR DE SERVIDOR");
+        localStorage.clear();
         if (this.oRouter !== undefined) {
           this.oRouter.navigate(['/home']);
         }
       }
       this.loading = false;
+      this._snackBar.dismiss();
     })
   }
 
@@ -72,36 +128,42 @@ export class ProductoRoutedPlistComponent implements OnInit {
     });
   }
 
-  nextPage(event: PageEvent) {
-    this.getPage(event.pageIndex, event.pageSize, this.currentSort, this.strToFind);
-  }
-
-  constructor(private productoService: ProductoService, private router: Router, private _location: Location, private _snackBar: MatSnackBar) {
-    this.oRouter = router;
-    this.getPage(this.pageIndex, this.pageSize, this.currentSort, this.strToFind);
+  constructor(private productoService: ProductoService, private oRouter: Router, private oActivatedRoute: ActivatedRoute, private _location: Location, private _snackBar: MatSnackBar) {
+    oActivatedRoute.params.subscribe(val => {
+      if (this.oActivatedRoute.snapshot.params.page) {
+        this.currentPageIndex = Number(this.oActivatedRoute.snapshot.params.page);
+        if ((this.currentPageIndex) <= 0) {
+          this.currentPageIndex = 1;
+        }
+      } else {
+        this.currentPageIndex = 1;
+      }
+      if (this.oActivatedRoute.snapshot.params.rpp) {
+        this.currentPageSize = Number(this.oActivatedRoute.snapshot.params.rpp);
+        if ((this.currentPageSize) >100) {
+          this.currentPageSize = 100;
+        }
+      } else {
+        this.currentPageSize = 10;
+      }
+      this.currentSortField = (this.oActivatedRoute.snapshot.params.sort);
+      this.currentSortDirection = (this.oActivatedRoute.snapshot.params.dir);
+      this.currentFilter = (this.oActivatedRoute.snapshot.params.filter);
+      this.getPage(this.currentPageIndex, this.currentPageSize, this.currentSortField, this.currentSortDirection, this.currentFilter);
+    });
   }
 
   ngOnInit() {
-  }
-
-  goToDetails(cod: number) {
-    this.router.navigate(['/producto-detail/', cod]);
   }
 
   goBack() {
     this._location.back();
   }
 
-  strToFind: string | undefined = undefined;
   onSubmitFindForm() {
-    //console.log(this.strToFind);
-    this.getPage(0, this.pageSize, this.currentSort, this.strToFind);
-  }
+    console.log("onSubmitFindForm", this.currentFilter);
+    this.oRouter.navigate(['/producto/plist/', this.currentPageIndex, this.currentPageSize, this.currentSortField, this.currentSortDirection, this.currentFilter]);
 
-  sortData(sort: Sort) {
-    console.log("sorting");
-    this.currentSort = sort;
-    this.getPage(this.pageIndex, this.pageSize, sort, this.strToFind);
   }
 
 }
